Abstracting out the github auth to its own file
