+++ src/api/start.js
@@ -2,5 +2,5 @@ const app = require('./app.js');
 const port = process.env.PORT || 3000;
 
 app.listen(port, () => {
-  console.log('Server running on http://localhost:' + port)
+  console.log('Server running on http://localhost:' + port);
 });
