+++ src/public/index.html
@@ -9,6 +9,6 @@
       <button type="submit">Login To Github</button>
     </form>
 
-    <script src="./build/bundle.js" type="text/js"></script>
+    <script src="./build/bundle.js" type="text/javascript"></script>
   </body>
 </html>
